Remove JMTestConfiguration and use JMSpecs instead
This is a refactor to remove JMTestConfiguration because it was serving a
similar purpose to JMSpec and we had been just copying between these two
data structures. JMSpec now contains the test configuration and has helper
methods to output this configuration as command line argument strings used
by JMeterRunner.

Beginning of converting some input properties from lists to maps as they
are representing key-value pairs.